refactor(ai): extract diplomatic candidate scoring (Refs #2289)
Move computeDiplomaticCandidateScores and cooldown helper into
diplomatic_candidate_scoring.dart so diplomacy_planner.dart stays
orchestration-only; barrel exports unchanged via diplomacy_planner re-exports.
